requirements.txt
Django==1.7
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.1.1
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
django-crispy-forms==1.4.0
